feat: parent config added
promos module added
