+++ microservice-promos/src/main/java/com/innovatech/promos/microservicepromos/MicroservicePromosApplication.java
@@ -1,0 +1,13 @@
+package com.innovatech.promos.microservicepromos;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MicroservicePromosApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MicroservicePromosApplication.class, args);
+	}
+
+}
