+++ microservice-promos/src/main/java/com/innovatech/promos/microservicepromos/promos/domain/model/queries/GetPromoByIdQuery.java
@@ -1,0 +1,12 @@
+package com.innovatech.promos.microservicepromos.promos.domain.model.queries;
+
+public record GetPromoByIdQuery(Long id) {
+    public GetPromoByIdQuery {
+        if (id == null) {
+            throw new IllegalArgumentException("id cannot be null");
+        }
+        if (id < 0) {
+            throw new IllegalArgumentException("id cannot be negative");
+        }
+    }
+}
