+++ microservice-promos/src/main/java/com/innovatech/promos/microservicepromos/promos/domain/model/queries/GetPromosQuery.java
@@ -1,0 +1,4 @@
+package com.innovatech.promos.microservicepromos.promos.domain.model.queries;
+
+public record GetPromosQuery() {
+}
