+++ microservice-promos/src/test/java/com/innovatech/promos/microservicepromos/MicroservicePromosApplicationTests.java
@@ -1,0 +1,13 @@
+package com.innovatech.promos.microservicepromos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MicroservicePromosApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
