--- src/main/java/pe/upc/trackmyroute/TrackMyRouteApplication.java
@@ -1,13 +1,0 @@
-package pe.upc.trackmyroute;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TrackMyRouteApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TrackMyRouteApplication.class, args);
-	}
-
-}
