--- src/main/java/pe/upc/trackmyroute/promos/domain/model/queries/GetPromosQuery.java
@@ -1,4 +1,0 @@
-package pe.upc.trackmyroute.promos.domain.model.queries;
-
-public record GetPromosQuery() {
-}
