--- src/main/java/pe/upc/trackmyroute/shared/domain/model/aggregate/AuditableAbstractAggregateRoot.java
@@ -1,31 +1,0 @@
-package pe.upc.trackmyroute.shared.domain.model.aggregate;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.domain.AbstractAggregateRoot;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.util.Date;
-
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-public class AuditableAbstractAggregateRoot<T extends AbstractAggregateRoot<T>> extends AbstractAggregateRoot<T> {
-
-    @Id
-    @Getter
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Getter
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    private Date createdAt;
-
-    @Getter
-    @LastModifiedDate
-    @Column(updatable = false)
-    private Date updatedAt;
-}
-
