--- src/main/resources/application.properties
@@ -1,14 +1,0 @@
-spring.application.name=track-my-route
-
-# Spring Datasource Configuration
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5432/upc
-spring.datasource.username=postgres
-spring.datasource.password=root
-
-# Spring Data JPA Configuration
-spring.jpa.show-sql=true
-
-# Spring Data JPA Hibernate Configuration
-spring.jpa.hibernate.ddl-auto=update
-
