--- src/test/java/pe/upc/trackmyroute/TrackMyRouteApplicationTests.java
@@ -1,13 +1,0 @@
-package pe.upc.trackmyroute;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TrackMyRouteApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
